:memo: added return type to docs
